backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydicom==2.4.4
anthropic>=0.78.0
pillow>=10.2.0
numpy>=1.26.0
requests>=2.28.0
